moved svm inside svm-js directory
